// src/reports-module/interfaces/detalle-capacitacion.interface.ts
export interface DetalleCapacitacion {
  NUMERO: number;
  DOCUMENTO: string;
  Codigo: string;
  COLABORADOR: string;
  VersionDoc: number;
  FechaEvaluacion: string;
  Semana: number;
  Lectura: string;
  Capacitacion: string;
  Evaluacion: string;
  Calificacion: number;
  FirmaFacilitador: string;
  FirmaColaborador: string;
  Estatus: string;
  Departamento: string;
  Depto: string;
  TipoDocumento: string;
  Tipo: string;
  DepartamentoColaborador: string;
  PuestoColaborador: string;
  ID_COLABORADOR: number;
  ID_PLAN: number;
  NOMBRE_PLAN: string;
  FECHA_APLICACION_PLAN: string;
}